Added error message dialog
Function _showingError() will popup an alert box that will stay
on the screen until the 'ok' button is pressed, or the screen
is pressed outside the alert box.

The function is called when the response of the 'api' is empty
(meaning that the user has no public repositories, it can be
tested by searching for the user 'aaa') or when the response is
null(meaning that no user within the given name was found).